fix: fix unruggableMemecoin launch indexer
